1. Удален класс ASound 2. Функция datatimeEx теперь перегружена и считает разность по модулю 3. Исправлен баг с /n и <br> в сообщениях 4. GetOstatok может получать остаток в float 5. Изменены правила построения правил для парсера бб кодов. 6. В небольшие структуры добавлен новый конструктор, позволяющий короче заполнить структуру 7. beta динамические смайлы

+++ ACore/amath.h
@@ -4,7 +4,7 @@
 #include <qmath.h>
 namespace ACore
 {
-int GetOstatok(float x,float k);
+float GetOstatok(float x,float k);
 double GradusToRadian(double grad);
 }
 #endif // AMATH
